fix unit prop and default it only in the tGrid body
